requirements.txt
fastapi
uvicorn
streamlit
pymupdf
faiss-cpu
sentence-transformers
scikit-learn
nltk
rouge-score
openai
pandas